[#540] fixed many things. removed old unnecessary files, split .h files into .h and .cpp, removed commands that are not currently implemented, like sending motor angles, using LED_BUILTIN instead of an additional #define for LED
