+++ robot/rover/internal_comms/src/LSSServoMotor.cpp
@@ -1,28 +1,4 @@
-/*
- * Written by William Wells October 01, 2021
- *
- * This is a simple library for interfacing with the communication protocol of
- * the two Lynxmotion Smart Servos used in the gripper of the arm. This was
- * designed to be compatible with the current arm code as of now and any code in
- * the future. As well as being compatible with different microcontrollers.
- */
-
-#ifndef LSSSERVOMOTOR_H
-#define LSSSERVOMOTOR_H
-
-#include <Arduino.h>
-
-class LSSServoMotor {
- public:
-  LSSServoMotor(HardwareSerial* serialPort);
-  void writeActionCommand(unsigned int servoId, const char* actionCommand,
-                          int actionValue);
-  void writeActionCommand(unsigned int servoId, const char* actionCommand);
-  char* writeQueryCommand(unsigned int servoId, const char* queryCommand);
-
- private:
-  HardwareSerial* ServosSerialBus;
-};
+#include "../include/LSSServoMotor.h"
 
 LSSServoMotor::LSSServoMotor(HardwareSerial* serialPort) {
   ServosSerialBus = serialPort;
@@ -75,7 +51,7 @@ void LSSServoMotor::writeActionCommand(unsigned int servoId,
  *
  * @param servoId
  * @param queryCommand
- * @return response to query command from servo
+ * @return response to query command from servo NOTE: you have to free this.
  */
 char* LSSServoMotor::writeQueryCommand(unsigned int servoId,
                                        const char* queryCommand) {
@@ -94,5 +70,3 @@ char* LSSServoMotor::writeQueryCommand(unsigned int servoId,
 
   return buff;
 }
-
-#endif
